+++ dranksforyou2api/views/liquors.py
@@ -22,7 +22,8 @@ class LiquorView(ViewSet):
         """Handle POST operations"""
         liquor = Liquor.objects.create(
            name=request.data["name"],
-           image=request.data["image"]
+           image=request.data["image"],
+           uid=request.user
        )
         serializer = LiquorSerializer(liquor)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -31,6 +32,9 @@ class LiquorView(ViewSet):
         """Handle PUT requests to update a liquor"""
         try:
             liquor = Liquor.objects.get(pk=pk)
+            if liquor.uid != request.user:
+                return Response({'message': 'You do not have permission to delete this beverage'}, status=status.HTTP_403_FORBIDDEN)
+            
             liquor.name = request.data.get("name", liquor.name)
             liquor.image = request.data.get("image", liquor.image)
             liquor.save()
@@ -46,6 +50,9 @@ class LiquorView(ViewSet):
         """Handle DELETE requests to delete a liquor"""
         try:
             liquor = Liquor.objects.get(pk=pk)
+            if liquor.uid != request.user:
+                return Response({'message': 'You do not have permission to delete this beverage'}, status=status.HTTP_403_FORBIDDEN)
+            
             liquor.delete()
             return Response({'message': 'Liquor deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
         except Liquor.DoesNotExist:
@@ -54,5 +61,5 @@ class LiquorView(ViewSet):
 class LiquorSerializer(serializers.ModelSerializer):
     class Meta:
         model = Liquor
-        fields = ['id', 'name', 'image']
+        fields = ['id', 'name', 'image', 'uid']
         depth = 1
